+++ components/checks/passive/grep/x_frame_options.rb
@@ -7,13 +7,12 @@
 =end
 
 # @author  Tasos Laskos <tasos.laskos@arachni-scanner.com>
-# @version 0.1.2
 class Arachni::Checks::XFrameOptions < Arachni::Check::Base
 
     def run
         return if audited?( page.parsed_url.host ) ||
             page.response.headers.empty? ||
-            page.response.headers['X-Frame-Options']
+            page.response.headers['X-Frame-Options'] || page.code != 200
         audited( page.parsed_url.host )
 
         log(
@@ -27,7 +26,7 @@ class Arachni::Checks::XFrameOptions < Arachni::Check::Base
             name:        'Missing X-Frame-Options header',
             description: %q{Checks the host for a missing `X-Frame-Options` header.},
             author:      'Tasos Laskos <tasos.laskos@arachni-scanner.com>',
-            version:     '0.1.2',
+            version:     '0.1.3',
             elements:    [ Element::Server ],
 
             issue:       {