+++ spec/support/servers/checks/passive/grep/x_frame_options.rb
@@ -7,3 +7,8 @@ end
 get '/safe' do
     headers 'X-Frame-Options' => 'DENY'
 end
+
+get '/non-200' do
+    headers 'X-Frame-Options' => ''
+    [404, 'Not found']
+end